packages/agentsociety2/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentsociety2"
version = "2.0.0" # change it for each release
description = "LLM-based agent society"
authors = [
    { name = "Jun Zhang", email = "zhangjun990222@gmail.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
]
dependencies = [
    "click>=8.0.0",
    "geojson>=3.1.0",
    "numpy>=1.20.0,<2.0.0",
    "aiosqlite>=0.20.0",
    "pycityproto>=2.4.5",
    "pydantic>=2.10.4",
    "pyproj>=3.6.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.2",
    "Shapely>=2.0.6",
    "json-repair>=0.46.2",
    "litellm>=1.77.7",
    "mcp[cli]>=1.13.1",
    "fastmcp>=2.12.2",
    "ruamel-yaml>=0.18.15",
    "prompt-toolkit>=3.0.51",
    "mem0ai>=1.0.0",
    "qdrant-client>=1.7.0",
    "aiohttp>=3.10.11",
    "stringcase>=1.2.0",
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.35.0",
    "mineru[core]>=1.0.0",
    "torch>=2.9.1",
    "doclayout-yolo>=0.0.4",
    "ultralytics>=8.3.243",
    "transformers>=4.57.3",
    "ftfy>=6.3.1",
    "dill>=0.4.0",
    "pyclipper>=1.4.0",
    "black>=25.12.0",
    "sqlmodel>=0.0.16",
    "greenlet>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/tsinghua-fib-lab/agentsociety"
Repository = "https://github.com/tsinghua-fib-lab/agentsociety.git"
Issues = "https://github.com/tsinghua-fib-lab/agentsociety/issues"

[tool.hatch.build.targets.wheel]
packages = ["agentsociety2"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

# 注意：Mirothinker Agent 是可选功能，不在默认依赖中
# 如需使用，请参考 MiroThinker/README.md 进行配置
